Move posixKnownPathsLocator.ts
