+++ src/test/pythonEnvironments/discovery/locators/posixKnownPathsLocator.unit.test.ts
@@ -7,7 +7,7 @@ import * as executablesAPI from '../../../../client/common/utils/exec';
 import { PythonEnvKind, PythonEnvSource } from '../../../../client/pythonEnvironments/base/info';
 import { BasicEnvInfo } from '../../../../client/pythonEnvironments/base/locator';
 import { getEnvs } from '../../../../client/pythonEnvironments/base/locatorUtils';
-import { PosixKnownPathsLocator } from '../../../../client/pythonEnvironments/discovery/locators/services/posixKnownPathsLocator';
+import { PosixKnownPathsLocator } from '../../../../client/pythonEnvironments/base/locators/lowLevel/posixKnownPathsLocator';
 import { createBasicEnv } from '../../base/common';
 import { TEST_LAYOUT_ROOT } from '../../common/commonTestConstants';
 import { assertBasicEnvsEqual } from './envTestUtils';
